requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.3=py37h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
astor=0.8.1=py37haa95532_0
async-timeout=4.0.2=py37haa95532_0
asynctest=0.13.0=py_0
attrs=22.1.0=py37haa95532_0
blas=1.0=mkl
blinker=1.4=py37haa95532_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2023.5.7=h56e8100_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py37h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37haa95532_0
colorama=0.4.6=py37haa95532_0
cryptography=39.0.1=py37h21b164f_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.11.0=pyhd3eb1b0_0
fftw=3.3.9=h2bbff1b_1
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=py37h2bbff1b_0
gast=0.2.2=py37_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py37hc60d5dd_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=3.7.0=py37h3de5c98_0
hdf5=1.10.6=h1756f20_1
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py37haa95532_0
importlib-metadata=4.11.3=py37haa95532_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_1
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=py37hd77b12b_0
krb5=1.19.4=h5b6d351_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.0=h6c2663c_2
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.1=py37haa95532_0
matplotlib=3.5.3=py37haa95532_0
matplotlib-base=3.5.3=py37hd77b12b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
multidict=6.0.2=py37h2bbff1b_0
munkres=1.1.4=py_0
numpy=1.21.5=py37h7a0a035_3
numpy-base=1.21.5=py37hca35cd5_3
oauthlib=3.2.1=py37haa95532_0
openssl=1.1.1t=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=22.0=py37haa95532_0
pcre=8.45=hd77b12b_0
pillow=9.4.0=py37hd77b12b_0
pip=22.3.1=py37haa95532_0
ply=3.11=py37_0
protobuf=3.20.3=py37hd77b12b_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.4.0=py37haa95532_0
pyopenssl=23.0.0=py37haa95532_0
pyparsing=3.0.9=py37haa95532_0
pyqt=5.15.7=py37hd77b12b_0
pyqt5-sip=12.11.0=py37hd77b12b_0
pysocks=1.7.1=py37_1
python=3.7.16=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
qt-main=5.15.2=he8e5bd7_8
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h2bbfb41_5
requests=2.28.1=py37haa95532_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.7.3=py37h7a0a035_2
setuptools=65.6.3=py37haa95532_0
sip=6.6.2=py37hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
tensorboard=2.8.0=py37haa95532_0
tensorboard-data-server=0.6.1=py37haa95532_0
tensorboard-plugin-wit=1.8.1=py37haa95532_0
tensorflow=1.15.0=gpu_py37hc3743a6_0
tensorflow-base=1.15.0=gpu_py37h1afeea4_0
tensorflow-estimator=1.15.1=pypi_0
tensorflow-gpu=1.15.0=h0d30ee6_0
termcolor=1.1.0=py37haa95532_1
tf-slim=1.0=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py37h2bbff1b_0
typing-extensions=4.4.0=py37haa95532_0
typing_extensions=4.4.0=py37haa95532_0
urllib3=1.26.14=py37haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=0.16.1=py_0
wheel=0.38.4=py37haa95532_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
wrapt=1.14.1=py37h2bbff1b_0
xz=5.4.2=h8cc25b3_0
yarl=1.8.1=py37h2bbff1b_0
zipp=3.11.0=py37haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
